Apply suggestions from code review

+++ src/mongocrypt-marking.c
@@ -830,7 +830,7 @@ _mongocrypt_fle2_placeholder_to_insert_update_ciphertextForRange (
    // g:= array<EdgeTokenSet>
    {
       BSON_ASSERT (placeholder->sparsity >= 0 &&
-                   placeholder->sparsity <= SIZE_MAX);
+                   (uint64_t) placeholder->sparsity <= (uint64_t) SIZE_MAX);
       edges = get_edges (&insertSpec, (size_t) placeholder->sparsity, status);
       if (!edges) {
          goto fail;
@@ -1242,7 +1242,7 @@ _mongocrypt_fle2_placeholder_to_find_ciphertextForRange (
       // g:= array<EdgeFindTokenSet>
       {
          BSON_ASSERT (placeholder->sparsity >= 0 &&
-                      placeholder->sparsity <= SIZE_MAX);
+                      (uint64_t) placeholder->sparsity <= (uint64_t) SIZE_MAX);
          mincover = mc_get_mincover_from_FLE2RangeFindSpec (
             &findSpec, (size_t) placeholder->sparsity, status);
          if (!mincover) {
